#43 add wxWidgets-3.0 support under Windows

+++ src/wxmedit/wxm_lines.h
@@ -302,7 +302,7 @@ struct MadLine
         while(true)
         {
             cnt=size;
-            if(wxFileOffset(pos+cnt) > biter->m_Size) cnt=biter->m_Size-pos;
+            if(wxFileOffset(pos+cnt) > biter->m_Size) cnt=(size_t)(biter->m_Size-pos);
             biter->Get(pos, buf, cnt);
             if((size-=cnt)==0) break;
